+++ packages/components/src/spectrum/picker/index.ts
@@ -1,3 +1,2 @@
 export * from './Picker';
-export * from './PickerModel';
 export * from './usePickerScrollOnOpen';
